Move movecraft to hook package
