+++ cannons-bukkit/src/main/java/at/pavlov/cannons/hooks/movecraft/listener/ProjectileImpactListener.java
@@ -1,7 +1,7 @@
-package at.pavlov.cannons.movecraft.listener;
+package at.pavlov.cannons.hooks.movecraft.listener;
 
 import at.pavlov.cannons.event.ProjectileImpactEvent;
-import at.pavlov.cannons.movecraft.damagetype.ProjectileImpactDamage;
+import at.pavlov.cannons.hooks.movecraft.damagetype.ProjectileImpactDamage;
 import net.countercraft.movecraft.combat.features.tracking.DamageRecord;
 import net.countercraft.movecraft.combat.features.tracking.events.CraftDamagedByEvent;
 import net.countercraft.movecraft.craft.Craft;
